refactor constants file to navItems
