+++ src/components/nav-new/NavBarNew.jsx
@@ -2,77 +2,8 @@ import { ChevronDown, Menu } from "lucide-react";
 import PropTypes from "prop-types";
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import CustomAnchor from "./CustomAnchor";
-import CustomLink from "./CustomLink";
-
-const navItems = [
-  {
-    label: "File",
-    submenu: [
-      {
-        url: <CustomAnchor url="https://www.example.com/" label="Open" />,
-      },
-      {
-        url: <CustomAnchor url="https://www.example.com/" label="Save" />,
-      },
-    ],
-  },
-  {
-    label: "Window",
-    submenu: [
-      {
-        url: <CustomLink url="setup" label="Experimental Setup" />,
-      },
-      {
-        url: <CustomLink url="instrument" label="Instrument" />,
-      },
-      {
-        url: <CustomLink url="spectra" label="Spectra" />,
-      },
-      {
-        url: <CustomAnchor url="https://www.example.com/" label="Find Peaks" />,
-      },
-    ],
-  },
-  {
-    label: "Measure",
-    submenu: [
-      {
-        url: (
-          <CustomAnchor
-            url="https://www.example.com/"
-            label="Collect Background"
-          />
-        ),
-      },
-      {
-        url: (
-          <CustomAnchor url="https://www.example.com/" label="Collect Sample" />
-        ),
-      },
-      {
-        url: (
-          <CustomAnchor url="https://www.example.com/" label="Cancel Scan" />
-        ),
-      },
-    ],
-  },
-  {
-    label: "Help",
-    submenu: [
-      {
-        dialog: true,
-        label: "Tutorial",
-        url: "tutorialDialog",
-      },
-      {
-        dialog: true,
-        label: "Welcome Dialog",
-        url: "welcomeDialog",
-      },
-    ],
-  },
-];
+import { navItems } from "../../constants/navItems.jsx";
+import CustomAnchor from "./CustomAnchor.jsx";
 
 function NavBar({ toggleWelcome, toggleTutorial }) {
   const [dropdown, setDropdown] = useState(null);
