--- src/constants.js
@@ -1,66 +1,0 @@
-export const navItems = [
-  {
-    label: "File",
-    submenu: [
-      {
-        label: "Open",
-        url: "https://www.example.com/",
-      },
-      {
-        label: "Save",
-        url: "https://www.example.com/",
-      },
-    ],
-  },
-  {
-    label: "Window",
-    submenu: [
-      {
-        label: "Experimental Setup",
-        route: "setup",
-      },
-      {
-        label: "Instrument",
-        route: "instrument",
-      },
-      {
-        label: "Spectra",
-        route: "spectra",
-      },
-      {
-        label: "Find Peaks",
-        url: "https://www.example.com/",
-      },
-    ],
-  },
-  {
-    label: "Measure",
-    submenu: [
-      {
-        label: "Collect Background",
-        url: "https://www.example.com/",
-      },
-      {
-        label: "Collect Sample",
-        url: "https://www.example.com/",
-      },
-      {
-        label: "Cancel Scan",
-        url: "https://www.example.com/",
-      },
-    ],
-  },
-  {
-    label: "Help",
-    submenu: [
-      {
-        label: "Tutorial",
-        url: "https://www.example.com/",
-      },
-      {
-        label: "Landing Page",
-        url: "https://www.example.com/",
-      },
-    ],
-  },
-];
